+++ Classes/ViewHelpers/TranslateViewHelper.php
@@ -19,12 +19,9 @@ use PSB\PsbFoundation\Utility\ContextUtility;
 use PSB\PsbFoundation\ViewHelpers\Translation\RegisterLanguageFileViewHelper;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\NotFoundExceptionInterface;
-use Psr\Http\Message\ServerRequestInterface;
 use TYPO3\CMS\Core\Configuration\Exception\ExtensionConfigurationExtensionNotConfiguredException;
 use TYPO3\CMS\Core\Configuration\Exception\ExtensionConfigurationPathDoesNotExistException;
-use TYPO3\CMS\Core\Http\ApplicationType;
-use TYPO3\CMS\Core\Localization\LanguageService;
-use TYPO3\CMS\Core\Localization\LanguageServiceFactory;
+use TYPO3\CMS\Core\Context\Exception\AspectNotFoundException;
 use TYPO3\CMS\Core\Utility\ArrayUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\Configuration\Exception\InvalidConfigurationTypeException;
@@ -36,13 +33,15 @@ use TYPO3Fluid\Fluid\Core\ViewHelper\AbstractViewHelper;
 use TYPO3Fluid\Fluid\Core\ViewHelper\Exception;
 use TYPO3Fluid\Fluid\Core\ViewHelper\Traits\CompileWithRenderStatic;
 use function count;
+use function in_array;
+use function is_array;
 
 /**
  * Class TranslateViewHelper
  *
  * Extended clone of the core ViewHelper in order to use \PSB\PsbFoundation\Service\LocalizationService which is able
  * to log missing language labels.
- * Additionally it provides a more conveniant way to pass variables into translations:
+ * Additionally, it provides a more convenient way to pass variables into translations:
  * Instead of:
  * <f:translate arguments="{0: 'myVar', 1: 123} id="myLabel" />
  * <source>My two variables are %1$s and %2$s.</source>
@@ -75,6 +74,7 @@ class TranslateViewHelper extends AbstractViewHelper
      * @param RenderingContextInterface $renderingContext
      *
      * @return string|null
+     * @throws AspectNotFoundException
      * @throws ContainerExceptionInterface
      * @throws ExtensionConfigurationExtensionNotConfiguredException
      * @throws ExtensionConfigurationPathDoesNotExistException
@@ -88,11 +88,13 @@ class TranslateViewHelper extends AbstractViewHelper
         RenderingContextInterface $renderingContext,
     ): ?string {
         [
-            'arguments'     => $translateArguments,
-            'default'       => $default,
-            'extensionName' => $extensionName,
-            'id'            => $id,
-            'key'           => $key,
+            'arguments'         => $translateArguments,
+            'default'           => $default,
+            'excludedLanguages' => $excludedLanguages,
+            'extensionName'     => $extensionName,
+            'id'                => $id,
+            'key'               => $key,
+            'languageKey'       => $languageKey,
         ] = $arguments;
 
         // Use key if id is empty.
@@ -108,6 +110,18 @@ class TranslateViewHelper extends AbstractViewHelper
 
         if ($renderingContext instanceof RenderingContext) {
             $request = $renderingContext->getRequest();
+
+            if (is_array($excludedLanguages)) {
+                $locale = $request?->getAttribute('language')->getLocale()->getName();
+
+                array_walk($excludedLanguages, static function (&$languageKey) {
+                    $languageKey = str_replace('_', '-', $languageKey);
+                });
+
+                if (in_array($locale, $excludedLanguages, true)) {
+                    return null;
+                }
+            }
         }
 
         if (!str_starts_with($id, FilePathUtility::LANGUAGE_LABEL_PREFIX)) {
@@ -122,7 +136,7 @@ class TranslateViewHelper extends AbstractViewHelper
         }
 
         try {
-            $value = static::translate($id, $extensionName, $translateArguments, $arguments['languageKey']);
+            $value = static::translate($id, $extensionName, $translateArguments, $languageKey);
         } catch (InvalidArgumentException) {
             $value = null;
         }
@@ -149,29 +163,6 @@ class TranslateViewHelper extends AbstractViewHelper
     }
 
     /**
-     * @param ServerRequestInterface|null $request
-     *
-     * @return LanguageService
-     * @see \TYPO3\CMS\Fluid\ViewHelpers\TranslateViewHelper
-     */
-    protected static function getLanguageService(ServerRequestInterface $request = null): LanguageService
-    {
-        if (isset($GLOBALS['LANG'])) {
-            return $GLOBALS['LANG'];
-        }
-        $languageServiceFactory = GeneralUtility::makeInstance(LanguageServiceFactory::class);
-        if ($request !== null && ApplicationType::fromRequest($request)->isFrontend()) {
-            $GLOBALS['LANG'] = $languageServiceFactory->createFromSiteLanguage($request->getAttribute('language')
-                ?? $request->getAttribute('site')->getDefaultLanguage());
-
-            return $GLOBALS['LANG'];
-        }
-        $GLOBALS['LANG'] = $languageServiceFactory->createFromUserPreferences($GLOBALS['BE_USER'] ?? null);
-
-        return $GLOBALS['LANG'];
-    }
-
-    /**
      * Wrapper call to static LocalizationService
      *
      * @param string      $id            Translation Key
@@ -180,9 +171,11 @@ class TranslateViewHelper extends AbstractViewHelper
      * @param string|null $languageKey   Language key to use for this translation
      *
      * @return string|null
+     * @throws AspectNotFoundException
      * @throws ContainerExceptionInterface
      * @throws ExtensionConfigurationExtensionNotConfiguredException
      * @throws ExtensionConfigurationPathDoesNotExistException
+     * @throws InvalidConfigurationTypeException
      * @throws JsonException
      * @throws NotFoundExceptionInterface
      */
@@ -252,6 +245,7 @@ class TranslateViewHelper extends AbstractViewHelper
         $this->registerArgument('arguments', 'array', 'Arguments to be replaced in the resulting string', false, []);
         $this->registerArgument('default', 'string',
             'If the given locallang key could not be found, this value is used. If this argument is not set, child nodes will be used to render the default');
+        $this->registerArgument('excludedLanguages', 'array', 'List of language keys that should return null');
         $this->registerArgument('extensionName', 'string', 'UpperCamelCased extension key (for example BlogExample)');
         $this->registerArgument('id', 'string', 'Translation ID. Same as key.');
         $this->registerArgument('key', 'string', 'Translation Key');